Move components into /auth directory.
