+++ src/app/auth/components/PreviousDomainListComponent.js
@@ -1,0 +1,26 @@
+import React from 'react';
+import propTypes from 'prop-types';
+
+const PreviousDomainListComponent = ({ previousDomains, switchLoginClick }) => {
+  return (
+    <div className="switch">
+      {previousDomains.map(addr => (
+        <button
+          type="button"
+          className="switchButton"
+          onClick={() => switchLoginClick(addr.domain)}
+        >
+          {addr.domain}
+        </button>
+      ))}
+    </div>
+  );
+};
+
+PreviousDomainListComponent.propTypes = {
+  // eslint-disable-next-line react/forbid-prop-types
+  previousDomains: propTypes.array.isRequired,
+  switchLoginClick: propTypes.func.isRequired,
+};
+
+export default PreviousDomainListComponent;
