+++ src/app/auth/components/index.js
@@ -1,3 +1,4 @@
 export { default as AuthModalComponent } from './AuthModalComponent';
 export { default as StartButtonComponent } from './StartButtonComponent';
 export { default as AuthTabComponent } from './AuthTabComponent';
+export { default as LoginDropdownComponent } from './LoginDropdownComponent';
