+++ src/app/auth/containers/index.js
@@ -1,3 +1,4 @@
 export { default as AuthModalContainer } from './AuthModalContainer';
 export { default as StartButtonContainer } from './StartButtonContainer';
 export { default as AuthTabContainer } from './AuthTabContainer';
+export { default as LoginDropdownContainer } from './LoginDropdownContainer';
