+++ src/app/components/index.js
@@ -17,4 +17,3 @@ export { default as AddressInputComponent } from './AddressInputComponent';
 export { default as UserErrorComponent } from './UserErrorComponent';
 export { default as UserSuccessComponent } from './UserSuccessComponent';
 export { default as UserWaitingComponent } from './UserWaitingComponent';
-export { default as LoginDropdownComponent } from './LoginDropdownComponent';
